# perf(ops): optimize permission check