Removes the overhead of permission check (logging) on access granted (common case):

Delta measured on `perf_now` from `deno_common` bench:

```
... Before ...
perf_now:            	n = 500000, dt = 0.264s, r = 1893939/s, t = 528ns/op
... After ...
perf_now:            	n = 500000, dt = 0.083s, r = 6024096/s, t = 166ns/op
```

So ~3x faster on `op_now` (which is a low-overhead op)

## Why is this faster ?

Avoids "expensive" `format` calls if no debug logging is enabled