@MierenManz 2021-08-21
@AaronO i think you may got the before and after values wrong. Unless it got about 3x slower 😅